+++ application/src/main/java/org/ehrbase/application/config/security/OAuth2SecurityConfiguration.java
@@ -17,7 +17,6 @@
  */
 package org.ehrbase.application.config.security;
 
-import static org.ehrbase.application.config.security.SecurityProperties.ADMIN;
 import static org.springframework.security.config.Customizer.withDefaults;
 
 import java.util.Arrays;
@@ -45,6 +44,7 @@ import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationConverter;
 import org.springframework.security.oauth2.server.resource.web.authentication.BearerTokenAuthenticationFilter;
 import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 
 /**
  * {@link Configuration} for OAuth2 authentication.
@@ -100,11 +100,11 @@ public class OAuth2SecurityConfiguration {
         http.cors(withDefaults())
                 .authorizeHttpRequests(auth -> {
                     AuthorizeHttpRequestsConfigurer<HttpSecurity>.AuthorizationManagerRequestMatcherRegistry registry =
-                            auth.requestMatchers("/rest/admin/**", "/management/**")
-                                    .hasRole(ADMIN);
+                            auth.requestMatchers(AntPathRequestMatcher.antMatcher("/rest/admin/**"))
+                                    .hasRole(adminRole);
 
-                    var managementAuthorizedUrl =
-                            registry.requestMatchers(this.managementWebEndpointProperties.getBasePath() + "/**");
+                    var managementAuthorizedUrl = registry.requestMatchers(AntPathRequestMatcher.antMatcher(
+                            this.managementWebEndpointProperties.getBasePath() + "/**"));
 
                     switch (managementEndpointsAccessType) {
                         case ADMIN_ONLY ->